+++ CHANGELOG.md
@@ -24,6 +24,7 @@
     - Removed the distinction between "visible" sets and "selected" (i.e. checked) sets.
 - Upgrade deck.gl to 8.3 and viv to 0.5
 - Fix build issues with nebula.
+- Use URL fragment instead of query param for `export-utils`.
 - Store updates to the grid layout (removing / moving / resizing) in the view config zustand store.
 - Clear the cell set selection when selecting a gene in `GenesSubsrciber`. Clear the gene selection when selecting cell set(s) in `CellSetsManagerSubscriber`.